more focus trapping, inert logic to mount functionality/inner wrapping component.
