+++ lib/Modal/Modal.js
@@ -1,20 +1,17 @@
-import React, { forwardRef, useRef, useEffect, useImperativeHandle, useCallback, useState } from 'react';
+import React, { forwardRef, useRef, useCallback, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { FormattedMessage } from 'react-intl';
-import * as focusTrap from 'focus-trap';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import { Transition, TransitionGroup } from 'react-transition-group';
 import { uniqueId, noop } from 'lodash';
-import useOverlayContainer from './useOverlayContainer';
-import { listen } from '../../util/listen';
 import { getFirstFocusable } from '../../util/getFocusableElements';
-import { MAIN_NAVIGATION_SELECTOR, OVERLAY_CONTAINER_SELECTOR, OVERLAY_CONTAINER_ID } from '../../util/consts';
+import useOverlayContainer from './useOverlayContainer';
+import WrappingElement from './WrappingElement';
+import { OVERLAY_CONTAINER_ID } from '../../util/consts';
 import IconButton from '../IconButton';
 import Headline from '../Headline';
 import css from './Modal.css';
-import calloutCSS from '../Callout/Callout.css';
-import overlayCSS from '../Popper/Popper.css';
 
 /**  getA11yLabel will return the appropriate labelling aria-attribute based on the props passed to the component
 * in HTML usage, if both "aria-label' is used, it will be applied. By default, the Modal will be labeled by its supplied
@@ -32,30 +29,6 @@ function getA11yLabel(props, labelId) {
     return { 'aria-labelledby': ariaLabelledbyProp };
   }
   return { 'aria-labelledby': labelId };
-}
-
-// applies/removes the `inert` attribute to elements outside of the `OverlayContainer` element.
-// `inert` blocks focus/mouse clicks as well as prevents screen readers from accessing those parts
-// of the document.
-export function useAssistiveTechTrap() {
-  const apply = () => {
-    const mainNav = document.querySelector(MAIN_NAVIGATION_SELECTOR); // main navigation
-    const overlaySiblings = Array.from(document.querySelectorAll(     // siblings of the overlay container
-      `${OVERLAY_CONTAINER_SELECTOR} ~ *`
-    ));
-    const elementList = [mainNav, ...overlaySiblings];
-    elementList.forEach(el => el?.setAttribute?.('inert', ''));
-  }
-
-  const release = () => {
-    const inertElements = Array.from(document.querySelectorAll('[inert]'));
-    inertElements?.forEach(el => el?.removeAttribute?.('inert'));
-  }
-
-  return {
-    release,
-    apply
-  }
 }
 
 const propTypes = {
@@ -130,46 +103,35 @@ const Modal = forwardRef((props, ref) => {
     scope = 'module',
     showHeader = true,
     size = 'medium',
-    wrappingElement = 'div',
+    wrappingElement: wrappingTagName = 'div',
     ...rest
   } = props;
   const modalId = useRef(id || uniqueId('modal')).current;
 
+  // Ref for the dialog element of the modal
+  const modalElem = useRef(null);
+
+  // Ref to the outer element provided as props but defaults as a div.
+  const wrappingElementRef = useRef(null);
+
   // get DOM element for portal. Within the FOLIO platform, this will be the `div#OverlayContainer`, but
   // component-level test suites, this could be `div#root`.
   const portalRef = useOverlayContainer(document.getElementById(OVERLAY_CONTAINER_ID));
-  // Ref for the `focus-trap` instance.
-  const focusTrapRef = useRef(null);
-  // Ref for the dialog element of the modal
-  const modalElem = useRef(null);
   // Ref for the element that was focused prior to the modal being opened so that we can return focus when the
   // modal closes.
   const triggerRef = useRef(null);
   // Ref to the element containing the modal content.
   const contentRef = useRef(null);
-  // Ref to the outer element provided as props but defaults as a div.
-  const wrappingElementRef = useRef(null);
-
-  // mechanism for applying 'inert' attribute (screen-reader trap);
-  const ATTrap = useAssistiveTechTrap();
-
-  // imperitive handle since we want to provide the wrapping element ref to the consuming app as the 'ref' prop.
-  useImperativeHandle(ref, () => wrappingElementRef.current, [wrappingElementRef]);
-
-  // refs to hold handler-removal functions so that any DOM-level listeners can be removed when the modal is hidden
-  // or dismounts.
-  const keyDownHandlerRef = useRef(null);
-  const focusHandlerRef = useRef(null);
-
-  // keydown Handler for detecting 'escape' key presses...
-  const keyDownHandler = useCallback((e) => {
-    if (e.key.toLowerCase() === 'escape') {
-      e.preventDefault();
-      onClose();
-    }
-  }, [onClose]);
 
   const [contentFocused, setContentFocused] = useState(false);
+
+  const refreshPortal = useCallback(() => {
+    if (portalRef.element) {
+      return portalRef.element;
+    }
+    portalRef.refresh(true);
+    return portalRef.element;
+  }, [portalRef.element]);
 
   // focus the first focusable element in the modal, falling back to the content
   // element itself if none is available...
@@ -184,117 +146,6 @@ const Modal = forwardRef((props, ref) => {
     }
   }, [wrappingElementRef]);
 
-  // if focus is outside of the modal or acceptable containers, return it to the modal.
-  const enforceFocusHandler = useCallback(() => {
-    // collect possible DOM elements for focus when a modal is active (<Modal>s, <Callout>s, <Popper>s (overlays)).
-    // check if the active element is contained among them to prevent multiple modals from fighting over focus.
-    const containers = Array.from(document.querySelectorAll(
-      `.${css.modalRoot}, .${calloutCSS.callout}, .${overlayCSS.overlay}`
-    ));
-    const focusIsWithin = containers.some((m) => m.contains(document.activeElement));
-
-    if (!focusIsWithin) {
-      sendFocusToModal();
-    }
-  }, [sendFocusToModal]);
-
-  const enforceFocusTimeout = useCallback(() => {
-    setTimeout(enforceFocusHandler);
-  }, [enforceFocusHandler]);
-
-  // mount/dismount effect - deactivate focus-trapping on dismount.
-  useEffect(() => {
-    return () => {
-      if (focusTrapRef.current?.active) {
-        focusTrapRef.current?.deactivate({
-          returnFocus: false,
-          checkCanReturnFocus: null
-        });
-      }
-      ATTrap.release();
-    }
-  }, []);
-
-  const refreshPortal = useCallback(() => {
-    if (portalRef.element) {
-      return portalRef.element;
-    }
-    portalRef.refresh(true);
-    return portalRef.element;
-  }, [portalRef.element]);
-
-  // open/closed effects - sending focus to modal, setting up focus-trapping,
-  // keyboard/focus listeners...
-  useEffect(() => {
-    if (open) {
-      const portal = refreshPortal();
-      if (modalElem.current &&
-        !triggerRef.current &&
-        !modalElem.current?.contains(document.activeElement)) {
-        triggerRef.current = document.activeElement;
-        sendFocusToModal();
-      }
-
-      if (modalElem.current) {
-        // create the focus trap if we don't already have one.
-        if (!focusTrapRef.current) {
-          focusTrapRef.current = focusTrap.createFocusTrap(portal, {
-            returnFocusOnDeactivate: false,
-            checkCanReturnFocus: null,
-            fallbackFocus: modalElem.current,
-            // we handle our own deactivation when the modal closes...
-            clickOutsideDeactivates: false,
-          });
-        }
-
-        if (enforceFocus) {
-          focusTrapRef.current.activate();
-          ATTrap.apply();
-        }
-        keyDownHandlerRef.current = listen(document, 'keydown', keyDownHandler);
-        focusHandlerRef.current = listen(document, 'focus', enforceFocusTimeout, true);
-      }
-    } else {
-      if (focusTrapRef.current?.active) {
-        focusTrapRef.current?.deactivate({
-          returnFocus: false,
-          checkCanReturnFocus: null
-        });
-        ATTrap.release();
-      }
-
-      // if lastActive is within the modal itself, focusManagement was probably handled
-      // outside of the component, so don't do anything if the suspected trigger is inside of the modal.
-      setContentFocused(false);
-      if (restoreFocus) triggerRef.current?.focus();
-
-      keyDownHandlerRef.current?.();
-      focusHandlerRef.current?.();
-    }
-
-    return () => {
-      if (focusTrapRef.current?.active) {
-        focusTrapRef.current?.deactivate({
-          returnFocus: false,
-          checkCanReturnFocus: null
-        });
-        focusTrapRef.current = null;
-        ATTrap.release();
-      }
-    }
-  }, [
-    open,
-    onClose,
-    onOpen,
-    restoreFocus,
-    enforceFocus,
-    sendFocusToModal,
-    enforceFocusTimeout,
-    keyDownHandler,
-    portalRef.element,
-    ATTrap,
-  ]);
-
   const handleBackgroundClick = useCallback(() => {
     if (!closeOnBackgroundClick) {
       // the modal isn't closing, so send focus back to the modal...
@@ -303,8 +154,6 @@ const Modal = forwardRef((props, ref) => {
     }
     onClose();
   }, [closeOnBackgroundClick, onClose, sendFocusToModal]);
-
-  const WrappingElement = wrappingElement;
 
   const getModalClass = () => classNames(
     css.modal,
@@ -325,10 +174,10 @@ const Modal = forwardRef((props, ref) => {
     }
   }, [contentFocused]);
 
-  if (!portalRef.element) {
-    refreshPortal();
-    return null;
-  }
+  // if (!portalRef.element) {
+  //   refreshPortal();
+  //   return null;
+  // }
 
   const modalElement = open ? (status) => createPortal(
     (
@@ -353,6 +202,16 @@ const Modal = forwardRef((props, ref) => {
             className={getModalClass()}
             id={props.id}
             ref={wrappingElementRef}
+            tagName={wrappingTagName}
+            triggerRef={triggerRef}
+            portalElement={portalRef.element}
+            modalElementRef={modalElem}
+            onClose={onClose}
+            onOpen={onOpen}
+            setContentFocused={setContentFocused}
+            enforceFocus={enforceFocus}
+            restoreFocus={restoreFocus}
+            sendFocusToModal={sendFocusToModal}
             {...rest}
           >
             {showHeader &&
